Initial commit: ft_printf data collection.
This commit contains a whole lot of code due to me coding it solo
before deciding to work on it as a team, it basically manages one
of the first steps in ft_printf, which is the collection of all
the information existing between every % sign and its conversion
specifier.

ft_printf will also use various functions from libft, so the libft
folder is also included, as well as the required Makefile,
the header file and finally the main file for testing purposes.

This is the starting point, please read all the comments carefully
to understand my approach, and make changes where you find
necessary, good luck.